Use request.view_args in the paginate macro()
Using request.view_args allows the macro to get any query string
args from the current context without needing to pass them to the
macro directly.

Request.endpoint was also added in the macro in url_for to eliminate
the need to pass a url function name to the macro.

The paginate macro now only takes a pagination object, and can be used
regardless of what page is using it.